+++ src/main/java/nextstep/auth/ui/AuthController.java
@@ -1,9 +1,9 @@
-package nextstep.member.ui;
+package nextstep.auth.ui;
 
 import lombok.RequiredArgsConstructor;
-import nextstep.member.application.AuthService;
-import nextstep.member.application.dto.TokenRequest;
-import nextstep.member.application.dto.TokenResponse;
+import nextstep.auth.application.AuthService;
+import nextstep.auth.application.dto.TokenRequest;
+import nextstep.auth.application.dto.TokenResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
